requirements.txt
appdirs==1.4.3
colorama==0.3.7
Django==1.10.6
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
django-crispy-forms==1.6.1
django-filter==1.0.2
djangorestframework==3.6.2
httpie==0.9.9
packaging==16.8
psycopg2==2.7.1
Pygments==2.2.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
